% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartel_duration.R
\name{add_nonlinears}
\alias{add_nonlinears}
\title{Nonlinear Variables}
\usage{
add_nonlinears(data, model)
}
\arguments{
\item{data}{cross-sectional cartel dataset}

\item{model}{int 1, 2, 3}
}
\value{
cross-sectional cartel dataset with nonlinear interdependencies
}
\description{
Function to add nonlinear interdependenciens
}
\examples{
sim_list <- sim_col_r();
df <- get_cartel_duration(sim_list$cartels_detected, sim_list$cartels_undetected, sim_list$interest_r, sim_list$deltas, sim_list$parms, model=1)
df <- add_nonlinears(df, model=1)
}
