% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartel_duration.R
\name{get_durations}
\alias{get_durations}
\title{Cartel Duration}
\usage{
get_durations(
  cartels_detected,
  cartels_undetected,
  r_all,
  delta_all,
  parms,
  model
)
}
\arguments{
\item{cartels_detected}{binary collusive state that eventually got detected}

\item{cartels_undetected}{binary collusive state that never got detected}

\item{r_all}{numeric stochastic interest rates}

\item{delta_all}{numeric discount factor based on interest rates}

\item{model}{int 1, 2, 3}
}
\value{
cross-sectional dataset with duration
}
\description{
Function to combine paneldata cartels and detected cartels returned by function sim_col_r into a cross-sectional dataset with duration
}
\examples{
sim_list <- sim_col_r();
df <- get_cartel_duration(sim_list$cartels_detected, sim_list$cartels_undetected, sim_list$interest_r, sim_list$deltas, sim_list$parms, model=1)
}
