% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartel_duration.R
\name{get_paneldata}
\alias{get_paneldata}
\title{Function to construct a panel dataset from sample and population of collusive and not collusive states over time}
\usage{
get_paneldata(cartels_population, cartels_detected, parms, r_all, sigma_all)
}
\arguments{
\item{cartels_population}{collusive and not collusive states of industries (population)}

\item{cartels_detected}{detected and undetected collusive and not collusive states of industries (sample)}

\item{parms}{parameter combinations used for simulation}

\item{interest_r}{stochastic values of interest rate over time}

\item{sigmas}{probability of detection over time}
}
\value{
panel dataset with cartels
}
\description{
Function to construct a panel dataset from sample and population of collusive and not collusive states over time
}
\examples{
sim_list <- sim_col_r();
df_panel <- get_paneldata(sim_list$cartels_population, sim_list$cartels_detected, sim_list$parms, sim_list$interest_r, sim_list$sigmas)
}
