% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_collusion_r.R
\name{sim_col_r}
\alias{sim_col_r}
\title{Simulate Collusion based on stochastic interest rates}
\usage{
sim_col_r(
  model = 1,
  periods = 1000,
  n_industries = 30,
  r_min = 0.001,
  r_max = 0.3,
  min_share = 1,
  n_firms_min = 2,
  n_firms_max = 20
)
}
\arguments{
\item{model}{Modelname (1, 2, 3)}

\item{periods}{Number of time periods}

\item{n_industries}{Number of industries}
}
\value{
List with detected cartels, undetected cartels, all cartels, discount factor over time, interest rate over time
}
\description{
Simulation of industries with collusive and non-collusive periods over time, detected and undetected, based on Stigler(1964) and inspired by Bos/Davies/Harrington/Ormosi(2018) and Harrington/Chang(2009)
}
\examples{
sim_list <- sim_col_r();
sim_list <- sim_col_r(model = "2", periods = 1000, n_ind = 20);
sim_list <- sim_col_r(model = "3", periods = 1000, n_ind = 30, n_firms_max=50);
}
